add LogDataSourceErrorAspect and @LogDataSourceError

+++ src/main/java/ru/t1/java/demo/dto/AccountDto.java
@@ -9,6 +9,7 @@ import lombok.NoArgsConstructor;
 import ru.t1.java.demo.model.AccountType;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 /**
  * DTO for {@link ru.t1.java.demo.model.Account}
@@ -20,11 +21,11 @@ import java.io.Serializable;
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class AccountDto implements Serializable {
     @JsonProperty("id")
-    private Integer id;
+    private Long id;
     @JsonProperty("client_id")
     private Long client_id;
     @JsonProperty("account_id")
     private AccountType accountType;
     @JsonProperty("balance")
-    private Double balance;
+    private BigDecimal balance;
 }
